Added network layer

+++ app/src/main/java/com/z80h3x/kezd_kov/data/orm/entities/Character.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class Character(
         @PrimaryKey(autoGenerate = true) val id : Long?,
-        @ColumnInfo(name = "cloudId") val cloudId: Int?,
+        @ColumnInfo(name = "cloudId") val cloudId: Long?,
         @ColumnInfo(name = "name") val name: String,
         @ColumnInfo(name = "description") val description: String?,
         @ColumnInfo(name = "initiative") val initiative: Int,
